backend/requirements.txt
pydantic==2.4.2
fastapi==0.104.0
uvicorn==0.23.2
motor==3.3.1
pydantic-settings==2.0.3
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
bcrypt==4.0.1
